Refactor datasets to be loaded without a buffer layer
